# lib/api-zod/src/generated/types/contradiction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuestPosition } from "./guestPosition";

export interface Contradiction {
  id: string;
  topic: string;
  canonical_topic: string;
  confidence: number;
  tension_summary: string;
  guest_a: GuestPosition;
  guest_b: GuestPosition;
}