# lib/api-zod/src/generated/types/guestPosition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuestPosition {
  name: string;
  position: string;
  quote: string;
  episode_title: string;
}