# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  total_contradictions: number;
  total_guests_featured: number;
  total_topics: number;
  highest_confidence_topic: string;
}